+++ src/lib/data/creatables/bsoItems.ts
@@ -5,13 +5,13 @@ import { expertCapesSource } from '../../bso/expertCapes';
 import { dyedItems } from '../../dyedItems';
 import { MaterialBank } from '../../invention/MaterialBank';
 import { nexBrokenArmorDetails } from '../../nex';
+import { allMasterCapesBank } from '../../skilling/skillcapes';
 import { bones } from '../../skilling/skills/prayer';
 import { Bone } from '../../skilling/types';
 import { assert, resolveNameBank, stringMatches } from '../../util';
 import getOSItem from '../../util/getOSItem';
 import itemID from '../../util/itemID';
 import resolveItems from '../../util/resolveItems';
-import { allMasterCapesBank } from '../buyables/bsoBuyables';
 import { brokenPernixOutfit, brokenTorvaOutfit, brokenVirtusOutfit } from '../CollectionsExport';
 import { Createable } from '../createables';
 import { slayerMaskCreatables } from './slayerMasks';
