+++ src/mahoji/commands/completion.ts
@@ -1,7 +1,8 @@
 import { AttachmentBuilder } from 'discord.js';
 import { ApplicationCommandOptionType, CommandRunOptions } from 'mahoji';
 
-import { calculateCompCapeProgress, generateAllCompCapeTasksList } from '../../lib/compCape';
+import { calculateCompCapeProgress } from '../../lib/bso/calculateCompCapeProgress';
+import { generateAllCompCapeTasksList } from '../../lib/compCape';
 import { OSBMahojiCommand } from '../lib/util';
 import { userStatsUpdate } from '../mahojiSettings';
